refactor: #359: make `pindexerDb` reusable
